refactor: consistently use lower case directory names
