+++ src/DiscoPoPExtension.ts
@@ -2,30 +2,30 @@ import * as vscode from 'vscode'
 import {
     Configuration,
     RunCapableConfiguration,
-} from './ConfigurationManager/Configuration'
-import configurationFromJSON from './ConfigurationManager/ConfigurationDeserializer'
-import { ConfigurationTreeDataProvider } from './ConfigurationManager/ConfigurationTreeDataProvider'
-import { Editable } from './ConfigurationManager/Editable'
-import { DiscoPoPCodeLensProvider } from './DiscoPoP/DiscoPoPCodeLensProvider'
-import { DiscoPoPDetailViewProvider } from './DiscoPoP/DiscoPoPDetailViewProvider'
-import { DiscoPoPParser } from './DiscoPoP/DiscoPoPParser'
+} from './configurationManager/Configuration'
+import configurationFromJSON from './configurationManager/ConfigurationDeserializer'
+import { ConfigurationTreeDataProvider } from './configurationManager/ConfigurationTreeDataProvider'
+import { Editable } from './configurationManager/Editable'
+import { DiscoPoPCodeLensProvider } from './discoPoP/DiscoPoPCodeLensProvider'
+import { DiscoPoPDetailViewProvider } from './discoPoP/DiscoPoPDetailViewProvider'
+import { DiscoPoPParser } from './discoPoP/DiscoPoPParser'
 import {
     DiscoPoPSuggestionGroup,
     DiscoPoPSuggestionNode,
     SuggestionTree,
-} from './DiscoPoP/DiscoPoPSuggestionTree'
-import { DiscoPoPResults } from './DiscoPoP/classes/DiscoPoPResults'
-import { Suggestion } from './DiscoPoP/classes/Suggestion/Suggestion'
-import { FileMapping } from './FileMapping/FileMapping'
-import { HotspotDetailViewProvider } from './HotspotDetection/HotspotDetailViewProvider'
-import { HotspotDetectionParser } from './HotspotDetection/HotspotDetectionParser'
-import { HotspotTree } from './HotspotDetection/HotspotTree'
-import { Hotspot } from './HotspotDetection/classes/Hotspot'
-import { HotspotDetectionResults } from './HotspotDetection/classes/HotspotDetectionResults'
-import { Commands } from './Utils/Commands'
-import { Decoration } from './Utils/Decorations'
-import { SimpleTreeNode } from './Utils/SimpleTree'
-import { UIPrompts } from './Utils/UIPrompts'
+} from './discoPoP/DiscoPoPSuggestionTree'
+import { DiscoPoPResults } from './discoPoP/classes/DiscoPoPResults'
+import { Suggestion } from './discoPoP/classes/Suggestion/Suggestion'
+import { FileMapping } from './fileMapping/FileMapping'
+import { HotspotDetailViewProvider } from './hotspotDetection/HotspotDetailViewProvider'
+import { HotspotDetectionParser } from './hotspotDetection/HotspotDetectionParser'
+import { HotspotTree } from './hotspotDetection/HotspotTree'
+import { Hotspot } from './hotspotDetection/classes/Hotspot'
+import { HotspotDetectionResults } from './hotspotDetection/classes/HotspotDetectionResults'
+import { Commands } from './utils/Commands'
+import { Decoration } from './utils/Decorations'
+import { SimpleTreeNode } from './utils/SimpleTree'
+import { UIPrompts } from './utils/UIPrompts'
 import { ToolSuite } from './runners/tools/ToolSuite'
 import { CancellationError } from './runners/helpers/cancellation/CancellationError'
