+++ src/configurationManager/ConfigurationTreeItem.ts
@@ -1,4 +1,4 @@
-import { SimpleTreeNode } from '../Utils/SimpleTree'
+import { SimpleTreeNode } from '../utils/SimpleTree'
 
 /** super class for everything that should be shown in the Configuration Tree */
 export interface ConfigurationTreeItem
