+++ src/configurationManager/Editable.ts
