+++ src/configurationManager/Property.ts
