+++ src/configurationManager/configurationImplementations/ConfigurationViewOnly.ts
