+++ src/discoPoP/classes/Suggestion/GenericSuggestion.ts
