+++ src/hotspotDetection/classes/Hotspot.ts
